inputModulus bad behavior
Fix numMin calculation -- old (wrong) behavior allowed numMax and numMin to both be nonzero. With current fix it should not be possible for numMax and numMin to both be nonzero. 

I do not know if this is mirrored in the C++ library.

An alternative one-liner with the native modulo operator is as follows:
`return ((input - minimumInput) % (maximumInput - minimumInput)) + minimumInput;`
This "zeros" the input, before wrapping it to the range with the modulo operator and finally undoing the offset.